+++ provisioning/terraform/00-terraform.tf
@@ -2,7 +2,32 @@
 # This file is used to define Terraform core resources
 
 terraform {
-  required_version = "~> 0.11"
+  # Required Terraform version
+  #
+  # Ensures the Terraform version used is compatible with this configuration.
+  #
+  # Source: https://www.terraform.io/language/settings#specifying-a-required-terraform-version
+  required_version = "~> 1.0"
+
+  # Required Terraform provider versions
+  #
+  # Terraform providers are versioned and distributed independently from Terraform itself.
+  #
+  # A registry maintains an index of 1st and vetted 3rd party providers. The `terraform init` command the latest
+  # version of each provider required here, subject to version constraints.
+  #
+  # Source: https://www.terraform.io/language/settings#specifying-provider-requirements
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 4.18"
+    }
+
+    azuread = {
+      source  = "hashicorp/azuread"
+      version = "~> 2.23"
+    }
+  }
 
   # AWS S3 Remote state backend
   #
@@ -13,7 +38,7 @@ terraform {
   # This resource relies on the AWS Terraform provider being previously configured.
   #
   # Source: https://gitlab.data.bas.ac.uk/WSF/terraform-remote-state
-  # Terraform source: https://www.terraform.io/docs/backends/types/s3.html
+  # Terraform source: https://www.terraform.io/language/settings/backends/s3
   backend "s3" {
     bucket = "bas-terraform-remote-state-prod"
     key    = "v2/BAS-ADD-DATA-CATALOGUE/terraform.tfstate"