+++ provisioning/terraform/58-acm_certificates.tf
@@ -13,12 +13,16 @@
 # This resource relies on the AWS Terraform provider ('us-east-1' alias) being previously configured
 #
 # AWS source: http://docs.aws.amazon.com/acm/latest/userguide/acm-overview.html
-# Terraform source: https://www.terraform.io/docs/providers/aws/r/acm_certificate.html
+# Terraform source: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/acm_certificate
 resource "aws_acm_certificate" "add-catalogue-integration" {
   provider = aws.us-east-1
 
   domain_name       = aws_s3_bucket.add-catalogue-integration.bucket
   validation_method = "DNS"
+
+  lifecycle {
+    create_before_destroy = true
+  }
 
   tags = {
     Name         = "add-catalogue-integration.data.bas.ac.uk"
@@ -40,6 +44,10 @@ resource "aws_acm_certificate" "add-catalogue-production" {
   domain_name       = aws_s3_bucket.add-catalogue-production.bucket
   validation_method = "DNS"
 
+  lifecycle {
+    create_before_destroy = true
+  }
+
   tags = {
     Name         = "add-catalogue.data.bas.ac.uk"
     X-Project    = "ADD Data Catalogue"
@@ -60,19 +68,26 @@ resource "aws_acm_certificate" "add-catalogue-production" {
 # This resource relies on the AWS Terraform provider being previously configured
 #
 # AWS source: http://docs.aws.amazon.com/Route53/latest/DeveloperGuide/rrsets-working-with.html
-# Terraform source: https://www.terraform.io/docs/providers/aws/r/route53_record.html
+# Terraform source: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/acm_certificate_validation
+# Terraform source: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/route53_record
 #
 # Tags are not supported by this resource
+//noinspection HILUnresolvedReference
 resource "aws_route53_record" "add-catalogue-integration-cert" {
-  zone_id = data.terraform_remote_state.BAS-CORE-DOMAINS.outputs.DATA-BAS-AC-UK-ID
+  for_each = {
+    for dvo in aws_acm_certificate.add-catalogue-integration.domain_validation_options : dvo.domain_name => {
+      name   = dvo.resource_record_name
+      record = dvo.resource_record_value
+      type   = dvo.resource_record_type
+    }
+  }
 
-  name = aws_acm_certificate.add-catalogue-integration.domain_validation_options.0.resource_record_name
-  type = aws_acm_certificate.add-catalogue-integration.domain_validation_options.0.resource_record_type
-  ttl  = 60
-
-  records = [
-    aws_acm_certificate.add-catalogue-integration.domain_validation_options.0.resource_record_value,
-  ]
+  allow_overwrite = true
+  name            = each.value.name
+  records         = [each.value.record]
+  ttl             = 60
+  type            = each.value.type
+  zone_id         = data.terraform_remote_state.BAS-CORE-DOMAINS.outputs.DATA-BAS-AC-UK-ID
 }
 
 # add-catalogue.data.bas.ac.uk
@@ -82,19 +97,26 @@ resource "aws_route53_record" "add-catalogue-integration-cert" {
 # This resource relies on the AWS Terraform provider being previously configured
 #
 # AWS source: http://docs.aws.amazon.com/Route53/latest/DeveloperGuide/rrsets-working-with.html
-# Terraform source: https://www.terraform.io/docs/providers/aws/r/route53_record.html
+# Terraform source: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/acm_certificate_validation
+# Terraform source: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/route53_record
 #
 # Tags are not supported by this resource
+//noinspection HILUnresolvedReference
 resource "aws_route53_record" "add-catalogue-production-cert" {
-  zone_id = data.terraform_remote_state.BAS-CORE-DOMAINS.outputs.DATA-BAS-AC-UK-ID
+  for_each = {
+    for dvo in aws_acm_certificate.add-catalogue-production.domain_validation_options : dvo.domain_name => {
+      name   = dvo.resource_record_name
+      record = dvo.resource_record_value
+      type   = dvo.resource_record_type
+    }
+  }
 
-  name = aws_acm_certificate.add-catalogue-production.domain_validation_options.0.resource_record_name
-  type = aws_acm_certificate.add-catalogue-production.domain_validation_options.0.resource_record_type
-  ttl  = 60
-
-  records = [
-    aws_acm_certificate.add-catalogue-production.domain_validation_options.0.resource_record_value,
-  ]
+  allow_overwrite = true
+  name            = each.value.name
+  records         = [each.value.record]
+  ttl             = 60
+  type            = each.value.type
+  zone_id         = data.terraform_remote_state.BAS-CORE-DOMAINS.outputs.DATA-BAS-AC-UK-ID
 }
 
 #    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *    *
@@ -112,14 +134,14 @@ resource "aws_route53_record" "add-catalogue-production-cert" {
 # This resource relies on the AWS Terraform provider ('us-east-1' alias) being previously configured
 #
 # AWS source: https://docs.aws.amazon.com/acm/latest/userguide/gs-acm-validate-dns.html
-# Terraform source: https://www.terraform.io/docs/providers/aws/r/acm_certificate_validation.html
+# Terraform source: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/acm_certificate_validation
 #
 # Tags are not supported by this resource
 resource "aws_acm_certificate_validation" "add-catalogue-integration" {
   provider = aws.us-east-1
 
   certificate_arn         = aws_acm_certificate.add-catalogue-integration.arn
-  validation_record_fqdns = [aws_route53_record.add-catalogue-integration-cert.fqdn]
+  validation_record_fqdns = [for record in aws_route53_record.add-catalogue-integration-cert : record.fqdn]
 }
 
 # add-catalogue.data.bas.ac.uk
@@ -131,12 +153,12 @@ resource "aws_acm_certificate_validation" "add-catalogue-integration" {
 # This resource relies on the AWS Terraform provider ('us-east-1' alias) being previously configured
 #
 # AWS source: https://docs.aws.amazon.com/acm/latest/userguide/gs-acm-validate-dns.html
-# Terraform source: https://www.terraform.io/docs/providers/aws/r/acm_certificate_validation.html
+# Terraform source: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/acm_certificate_validation
 #
 # Tags are not supported by this resource
 resource "aws_acm_certificate_validation" "add-catalogue-production" {
   provider = aws.us-east-1
 
   certificate_arn         = aws_acm_certificate.add-catalogue-production.arn
-  validation_record_fqdns = [aws_route53_record.add-catalogue-production-cert.fqdn]
+  validation_record_fqdns = [for record in aws_route53_record.add-catalogue-production-cert : record.fqdn]
 }